pyproject.toml
[tool.poetry]
name = "antibot-browser"
version = "0.1.0"
description = ""
authors = ["songponlekpetch <songpon.l@wisesight.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
playwright = "^1.50.0"
playwright-stealth = "^1.0.6"
shopee-captcha-solver = "^0.1.5"
chrome-extension-python = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
